Workaround for Dashboard 404 redirect bug
Since Jira 7.2.0 there has been an issue with redirecting to /secure/Dashboard.jspa

According to this upsteam Jira: https://jira.atlassian.com/browse/JRA-63278
it is related to redirecting to the dashboard URL with an absolute URL,
which this plugin does.

Not using absolute URLs seems to result to having to login twice, which
is not desirable. So the best solution at the moment is to filter a
redirect to: /secure/Dashboard.jspa with /default.jsp

Note: This commit depends on the following pull request being merged in atlassian-saml:
https://github.com/bitium/atlassian-saml/pull/7